L2CL usage for half-cycle ambiguity resolution speed-up (#398)
* Implement L2CL tracker

* Implement coherent phase discriminator

* Change code_phase to double

* Implement L2CM to L2CL handover

* Start L2CL only if L2CM ambiquity unknown

* L2CL management

* Address review comments

* Clean up L2CL decoder + rebase

* Fix bug with L2CL code phase

* Address review comments - misc_info

* Address review comments - fll_mode

* Address review comments - break loops

* Implement faster L2CL startup

* Drop unstable L2CL trackers quickly

* Phase comparison refactor

* Address review comments

* Remove L2CL data decoder

* Copyright updates

* Moved check_decoder_need function

* Move tracker specific code from track.c

* Search tracking channel based on sid

* Move L2CL data to common_data

* Move data decoder requirements to LSNP

* LSNP pointer

+++ src/board/nap/track_channel.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2011-2014,2016 Swift Navigation Inc.
+ * Copyright (C) 2011 - 2017 Swift Navigation Inc.
  * Contact: Fergus Noble <fergus@swift-nav.com>
  *          Colin Beighley <colin@swift-nav.com>
  *
@@ -27,7 +27,7 @@ extern u8 nap_track_n_channels;
 /** \} */
 
 void nap_track_init(u8 channel, gnss_signal_t sid, u32 ref_timing_count,
-                    float carrier_freq, float code_phase, u32 chips_to_correlate);
+                    float carrier_freq, double code_phase, u32 chips_to_correlate);
 
 void nap_track_update(u8 channel, double carrier_freq,
                       double code_phase_rate, u32 chips_to_correlate,
